バリデーションロジックを分離し、FoodFormから利用するように変更
