+++ src/components/FoodForm.tsx
@@ -26,45 +26,11 @@ import {
 } from '@/lib/date-utils';
 import { addFoodItem } from '@/lib/storage';
 import { Loader2 } from 'lucide-react';
-
-/**
- * 食品名の最大文字数
- */
-const MAX_NAME_LENGTH = 50;
-
-/**
- * HTMLタグを削除する関数
- * @param input 入力文字列
- * @returns サニタイズされた文字列
- */
-const sanitizeHtmlTags = (input: string): string => {
-  // <script> タグやその他のHTMLタグを削除
-  const noScriptTags = input.replace(
-    /<script\b[^<]*(?:(?!<\/script>)<[^<]*)*<\/script>/gi,
-    ''
-  );
-  // 残りの一般的なHTMLタグを削除
-  return noScriptTags.replace(/<\/?[^>]+(>|$)/g, '');
-};
-
-/**
- * 食品名のバリデーションを行う関数
- * @param name 食品名
- * @returns エラーメッセージ（エラーがない場合はnull）
- */
-const validateFoodName = (name: string): string | null => {
-  const trimmedName = name.trim();
-
-  if (!trimmedName) {
-    return '食品名を入力してください。空白文字のみは無効です。';
-  }
-
-  if (trimmedName.length > MAX_NAME_LENGTH) {
-    return `食品名は${MAX_NAME_LENGTH}文字以内で入力してください。`;
-  }
-
-  return null;
-};
+import {
+  MAX_NAME_LENGTH,
+  sanitizeHtmlTags,
+  validateFoodName,
+} from '@/lib/validation';
 
 interface FoodFormProps {
   /**
